apps/video-consumer/requirements.txt
confluent-kafka==2.0.2
pyyaml==6.0
opencv-python-headless==4.7.0.72
tensorflow==2.11.0